requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-bootstrap3==9.0.0
django-webpack-loader==0.5.0
djangorestframework==3.6.4
Faker==0.8.4
idna==2.6
ipaddress==1.0.18
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
pprint==0.1
psycopg2==2.7.3.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
uszipcode==0.1.3
virtualenv==15.1.0
